Improved logging

+++ UpdatedModel/ModelCode/CompleteModelCall.py
@@ -12,6 +12,7 @@ import sys
 import time as tm 
 import numpy as np
 from datetime import datetime
+from termcolor import colored
 
 from ModelCode.SettingsParameters import DefaultSettingsExcept
 from ModelCode.Auxiliary import filename
@@ -24,7 +25,7 @@ from ModelCode.GetPenalties import GetPenalties
 from ModelCode.ModelCore import SolveReducedcLinearProblemGurobiPy
 from ModelCode.VSSandValidation import VSS
 from ModelCode.VSSandValidation import OutOfSampleVal
-
+from ModelCode.GeneralSettings import logs_on
 
 # %% ############## WRAPPING FUNCTIONS FOR FOOD SECURITY MODEL ################
 
@@ -141,18 +142,26 @@ def FoodSecurityProblem(PenMet = "prob", probF = 0.99, probS = 0.95, \
     
     # if model output already exists, it is loaded
     if not os.path.isfile("ModelOutput/SavedRuns/" + fn + ".txt"):
-        crop_alloc, meta_sol, status, durations, settings, args, \
-        yield_information, population_information, rhoF, rhoS, VSS_value, \
-        crop_alloc_vss, meta_sol_vss, validation_values = \
-                            OptimizeModel(PenMet = PenMet,  
-                                          probF = probF, 
-                                          probS = probS, 
-                                          rhoFini = rhoF,
-                                          rhoSini = rhoS,
-                                          prints = prints,
-                                          validation = validation,
-                                          save = save,
-                                          **kwargs)
+        try:
+            crop_alloc, meta_sol, status, durations, settings, args, \
+            yield_information, population_information, rhoF, rhoS, VSS_value, \
+            crop_alloc_vss, meta_sol_vss, validation_values = \
+                                OptimizeModel(PenMet = PenMet,  
+                                              probF = probF, 
+                                              probS = probS, 
+                                              rhoFini = rhoF,
+                                              rhoSini = rhoS,
+                                              prints = prints,
+                                              validation = validation,
+                                              save = save,
+                                              **kwargs)
+        except KeyboardInterrupt:
+            print(colored("\nThe model run was interupted by the user.", "red"), flush = True)
+            if logs_on:
+                log = open("ModelLogs/tmp.txt", "a")
+                log.write("\n\nThe model run was interupted by the user.")
+                log.close()
+                os.rename("ModelLogs/tmp.txt", "ModelLogs/" + fn + ".txt")
         
     # if not, it is calculated
     else:            
@@ -275,22 +284,22 @@ def OptimizeModel(PenMet = "prob", probF = 0.99, probS = 0.95, \
     # timing
     all_start  = tm.time()
     
-    # TODO include decision for logfile in all printing statements so console
     # initialize log file
-    if os.path.exists("ModelLogs/tmp.txt"):
-        os.remove("ModelLogs/tmp.txt")
-    log = open("ModelLogs/tmp.txt", "a")
-    log.write("Model started " + str(datetime.now().strftime("%B %d, %Y, at %H:%M")))
-    log.write("\n\nModel Input: PenMet = " + str(PenMet) + 
-              "\n             probF = " + str(probF) + 
-              "\n             probS = " + str(probS) + 
-              "\n             rhoF = " + str(rhoFini) + 
-              "\n             rhoS = " + str(rhoSini) + 
-              "\n             validation = " + str(validation) + 
-              "\n             save = " + str(save))
-    for key in kwargs.keys():
-        log.write("\n             " + key + " = " + str(kwargs[key]))
-    log.close()
+    if logs_on:
+        if os.path.exists("ModelLogs/tmp.txt"):
+            os.remove("ModelLogs/tmp.txt")
+        log = open("ModelLogs/tmp.txt", "a")
+        log.write("Model started " + str(datetime.now().strftime("%B %d, %Y, at %H:%M")))
+        log.write("\n\nModel Input: PenMet = " + str(PenMet) + 
+                  "\n             probF = " + str(probF) + 
+                  "\n             probS = " + str(probS) + 
+                  "\n             rhoF = " + str(rhoFini) + 
+                  "\n             rhoS = " + str(rhoSini) + 
+                  "\n             validation = " + str(validation) + 
+                  "\n             save = " + str(save))
+        for key in kwargs.keys():
+            log.write("\n             " + key + " = " + str(kwargs[key]))
+        log.close()
         
     # create dictionary of all settings (includes calculating or loading the
     # correct expected income)
@@ -366,7 +375,8 @@ def OptimizeModel(PenMet = "prob", probF = 0.99, probS = 0.95, \
                 pickle.dump(probS, fp)
      
     # rename the temporal log file
-    os.rename("ModelLogs/tmp.txt", "ModelLogs/" + fn + ".txt")    
+    if logs_on:
+        os.rename("ModelLogs/tmp.txt", "ModelLogs/" + fn + ".txt")    
      
     # timing
     all_end  = tm.time()
